+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Add support for PHP v8.0 .
 - Type declarations have been added to all method parameters and return types
   where possible.
 ### Changed